% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGEscore.R
\name{autocor}
\alias{autocor}
\title{Computation of the autocorrelation}
\usage{
autocor(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
the autocorrelation.
}
\description{
The function autocor computes the autocorrelation.(function from localscore)
}
