% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGEtest.R
\name{CheckContrast}
\alias{CheckContrast}
\title{Check and reformat the matrix of contrast}
\usage{
CheckContrast(Contrast, ContrastName)
}
\arguments{
\item{Contrast}{A matrix of contrast.}

\item{ContrastName}{The name of the contrast.}
}
\value{
The matrix of contrast in the right format.
}
\description{
The function CheckContrast check and reformat the matrix of  contrast.
}
