% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGEtest.R
\name{CheckIncidence}
\alias{CheckIncidence}
\title{Check and reformat the matrix of incidence}
\usage{
CheckIncidence(Incidence, IncidenceName)
}
\arguments{
\item{Incidence}{A matrix of incidence, as obtained from \code{\link[=metaGE.incidence]{metaGE.incidence()}}.}

\item{IncidenceName}{The name of the incidence.}
}
\value{
The matrix of incidence in the right format.
}
\description{
The function CheckIncidence check and reformat the matrix of incidence.
}
