% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGEtest.R
\name{ContrastStatTest.NA}
\alias{ContrastStatTest.NA}
\title{Compute the statistic of the contrast test in presence of missing values}
\usage{
ContrastStatTest.NA(
  Incidence,
  Contrast = NULL,
  Zmat,
  MatCorr,
  Data,
  Configs.list,
  IncidenceName
)
}
\arguments{
\item{Incidence}{A matrix of incidence, as obtained from metaGE.incidence.}

\item{Contrast}{A matrix of contrast, if NULL the identity matrix is used. (\code{NULL} by default)}

\item{Zmat}{A matrix containing the Zscores of all markers (in rows) in each environment (in columns).}

\item{MatCorr}{The inter-environments correlation matrix. Can be computed using \code{\link[=metaGE.cor]{metaGE.cor()}}.}

\item{Data}{A dataset containing the effect, the pvalues and the na configuration for all marker}

\item{Configs.list}{A vector containing the \code{NA} configurations present in the dataset}

\item{IncidenceName}{The name of the incidence.}
}
\value{
A dataset of two columns containing the pvalue of the test of contrast and the minimum number of environment per group of all markers.
}
\description{
The function ContrastStatTest compute the statistic of the contrast test.
}
