% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGEtest.R
\name{ContrastStatTest}
\alias{ContrastStatTest}
\title{Compute the statistic of the contrast test.}
\usage{
ContrastStatTest(Incidence, Contrast = NULL, Zmat, MatCorr, IncidenceName)
}
\arguments{
\item{Incidence}{A matrix of incidence, as obtained from \code{\link[=metaGE.incidence]{metaGE.incidence()}}.}

\item{Contrast}{A matrix of contrast, if \code{NULL} the identity matrix is used. (\code{NULL} by default)}

\item{Zmat}{A matrix containing the Zscores of all markers (in rows) in each environment (in columns).}

\item{MatCorr}{The inter-environments correlation matrix. Can be computed using \code{\link[=metaGE.cor]{metaGE.cor()}}.}

\item{IncidenceName}{The name of the incidence.}
}
\value{
A dataset of two columns containing the pvalue of the test of contrast and the minimum number of environment per group of all markers.
}
\description{
The function ContrastStatTest compute the statistic of the contrast test.
}
