% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{envDesc}
\alias{envDesc}
\title{Description of the environments.}
\format{
A data frame with 22 rows and 3 variables:
\itemize{
\item FileName: environment name
\item Temp: temperature
\item Water: water condition
}
}
\usage{
envDesc
}
\description{
A dataset containing variables describing the 22 environments.
}
\keyword{datasets}
