% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGEmatcor.R
\name{GetH0Items}
\alias{GetH0Items}
\title{GetH0Items}
\usage{
GetH0Items(Zmat, Threshold = 0.8, plotting = FALSE, Cores = NULL)
}
\arguments{
\item{Zmat}{A matrix containing the Zscore (in rows) for each environment (in columns)}

\item{Threshold}{Threshold on posteriors (to be H1) to filter markers for correlation computation (0.6 by default)}

\item{plotting}{A boolean saying to plot or not (FALSE by default)}

\item{Cores}{The number of cores to used, optional. By default, availableCores()-1 cores is used.}
}
\value{
A vector of index of markers which seems not significant (under H0)
}
\description{
This function give the index of the markers which seems not significant (under H0)
}
