% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGEscore.R
\name{lindley}
\alias{lindley}
\title{Computation of the lindley process from scores.}
\usage{
lindley(scores)
}
\arguments{
\item{scores}{A numeric vector.}
}
\value{
the lindley.
}
\description{
The function lindley computes the lindley process from scores.(function from localscore)
}
