% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGEfit.R
\name{LLikelihoodT_vect}
\alias{LLikelihoodT_vect}
\title{LLikelihoodT_vect}
\usage{
LLikelihoodT_vect(Zmat, Delta, P, Mu, Tau)
}
\arguments{
\item{Zmat}{A matrix containing the Zscores of all markers (in rows) in each environment (in columns)}

\item{Delta}{A vector containing the diagonal coefficients of the diagonal matrix obtained by the diagonalization of the correlation matrix}

\item{P}{Matrix such that MatCorr = P Delta t(P), with Delta diagonal}

\item{Mu}{A vector containing the average effect of the markers}

\item{Tau}{A vector containing the heterogeneity between environments of the markers}
}
\value{
A vector containing the value of the Log-Likelihood of all markers
}
\description{
This function compute the values of loglikelihood for all markers.
}
