% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGEplot.R
\name{MakeQQplot}
\alias{MakeQQplot}
\title{Drawing a QQplot}
\usage{
MakeQQplot(Pvalues, Name = NULL, Xrange = NULL, Yrange = NULL)
}
\arguments{
\item{Pvalues}{A vector containing pvalues.}

\item{Name}{A name of the corresponding test. (optional)}

\item{Xrange}{A range for the x axis. (optional)}

\item{Yrange}{A range for the y axis. (optional)}
}
\description{
The function MakeQQplot displays the QQplot of the -log10(pvalues).
}
