% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metaData}
\alias{metaData}
\title{Results of different GWAS.}
\format{
A data frame with 25,436 rows and 35 variables:
\itemize{
\item CHR: chromosome of the marker
\item POS: position of the marker
\item MARKER: name of the marker
\item FREQ.env: maf of the marker in the environment env
\item EFFECT.env: regression coefficient of the marker in the environment env
\item EFFECT_SE.env: standard error of the regression coefficient of the marker in the environment env
\item PVAL.env: pvalue of the marker in the environment env
\item WEIGHT.env: weight of the marker in the environment env
\item ALLELE0: allele0
\item ALLELE1: allele1
}
}
\usage{
metaData
}
\description{
A dataset containing the results of 10 different genetic association studies testing the association between a set of 25,436 markers and the grain yield.
The data are extrated from: Drops Amaizing available on the https://doi.org/10.15454/6TL2N4 website.
This dataset were obtained thanks to the \emph{metaGE.collect} function.
}
\keyword{datasets}
