% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGEcollect.R
\name{ReadData}
\alias{ReadData}
\title{ReadData}
\usage{
ReadData(ListN, FileN, VarN, MinFreq = 0)
}
\arguments{
\item{ListN}{The name of the list of files where the file to read belongs or NULL if there is only one list of files}

\item{FileN}{The name of the file to read}

\item{VarN}{A named list containing the column names in the file corresponding to the variables below: MARKER, CHR, POS, EFFECT, PVAL. (optional: FREQ, ALLELE0, ALLELE1)}

\item{MinFreq}{A numeric value allowing to filter to keep markers with MAF > MinFreq}
}
\value{
A tibble with the interesting columns selected and renamed.
}
\description{
This function read the one file, select interesting columns and rename them.
}
