% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGEtest.R
\name{RegressionStatTest}
\alias{RegressionStatTest}
\title{Compute the pvalue of the meta-regression test.}
\usage{
RegressionStatTest(Covariate, CovName, Zmat, MatCorr)
}
\arguments{
\item{Covariate}{A dataset containing the values of one Covariate (in columns) in each environment (in rows).}

\item{CovName}{The name the Covariate.}

\item{Zmat}{A matrix containing the Zscores of all markers (in rows) in each environment (in columns).}

\item{MatCorr}{The inter-environments correlation matrix. Can be computed using \code{\link[=metaGE.cor]{metaGE.cor()}}.}
}
\value{
A dataset of two columns containing the pvalue of the meta-regression test and the number of environment used to perform the test of all markers.
}
\description{
The function RegressionStatTest compute the statistic and the pvalue of the regression test.
}
