% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGEtest.R
\name{RegressionStatTestNA}
\alias{RegressionStatTestNA}
\title{Compute the pvalue of the regression test in presence of missing values.}
\usage{
RegressionStatTestNA(Covariate, CovName, Zmat, MatCorr, Data, Configs.list)
}
\arguments{
\item{Covariate}{A dataset containing the values of one covariate (in columns) in each environment (in rows).}

\item{CovName}{The name the covariate.}

\item{Zmat}{A matrix containing the Zscores of all markers (in rows) in each environment (in columns).}

\item{MatCorr}{The inter-environments correlation matrix. Can be computed using \code{\link[=metaGE.cor]{metaGE.cor()}}.}

\item{Data}{A dataset containing the effect, the pvalues and the \code{NA} configuration for all marker}

\item{Configs.list}{A vector containing the NA configurations present in the dataset}
}
\value{
A dataset of two columns containing the pvalue of the meta-regression test and the number of environment used to perform the test of all markers.
}
\description{
The function RegressionStatTest compute the statistic and the pvalue of the regression test.
}
