% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGEscore.R
\name{sig_sl}
\alias{sig_sl}
\title{Computation of the significative regions}
\usage{
sig_sl(lind, pos, th)
}
\arguments{
\item{lind}{The lindley}

\item{pos}{The position}

\item{th}{The threshold}
}
\value{
the significance threshold.
}
\description{
The function sig_sl computes the significative regions from a lindley process given a significance threshold.(function from localscore)
}
