% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGEscore.R
\name{thresUnif}
\alias{thresUnif}
\title{Computation of the significance threshold}
\usage{
thresUnif(L, cor, xi, alpha = 0.05)
}
\arguments{
\item{L}{The length of the chromosome}

\item{cor}{The autocorrelation of the chromosome}

\item{xi}{The threshold of the score, xi = 1,2,3 or 4.}

\item{alpha}{The nominal threshold.}
}
\value{
the significance threshold.
}
\description{
The function thresUnif computes the significance threshold.(function from localscore)
}
\details{
The distribution of the p-values is uniform, the local score follows a Gumbel distribution under the null.
}
