% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGE-package.R
\docType{package}
\name{metaGE-package}
\alias{metaGE}
\alias{metaGE-package}
\title{metaGE: Meta-Analysis for Detecting Genotype x Environment Associations}
\description{
Meta-analysis of genome-wide association studies for studying Genotype x Environment interactions. The 4 main functions of the package metaGE.collect(), metaGE.cor(), metaGE.fit() and metaGE.test() correspond to 4 steps to perform the meta-analysis: Collecting the results of genome-wide association studies data from different files; Inferring the inter-environment correlation matrix; Performing global test procedure for quantitative trait loci detection (using a Fixed or Random effect model); Performing tests of contrast or meta-regression using an environmental co-factor. (De Walsche, A., et al. (2023) \doi{10.1101/2023.03.01.530237}).
}
\author{
\strong{Maintainer}: Annaïg De Walsche \email{annaig.de-walsche@inrae.fr} (\href{https://orcid.org/0000-0003-0603-1716}{ORCID})

Authors:
\itemize{
  \item Tristan Mary-Huard \email{tristan.mary-huard@agroparistech.fr} (\href{https://orcid.org/0000-0002-3839-9067}{ORCID})
}

}
\keyword{internal}
